chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-custom-element-cycle-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: document.registerElement is not a function
CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This is a testharness.js-based test.
FAIL import-custom-element-cycle Uncaught TypeError: document.registerElement is not a function
Harness: the test ran to completion.
